Move onboarding into its own folder
